Fix digit tick animation cancellation
Previously if you were to reset the timer while the digit tick animation was playing, the digits would be offset since they were preserving the animation values creating an unintended visual bug. The ResetTextPosition method resets the position of the text so the animation values don't get saved / carried over.