Move all digit logic into digit format class
We've changed the flow and layout a bit so it'd be wise to take another look at this from top to bottom to understand how the application is currently running. Almost done with the refactor, have a couple bugs to still iron out but should be working for the most part using different formats too! :D