Add interruptibility parameter to timer confirmation dialogs
On timer completion, we will close our active dialog if one is present. However with the new Restart Application action, we don't want to hide our popup since it's still relevant regardless of what the timer does/is. This allows us to keep open the confirmation dialog when prompted to restart settings to their factory settings.