Fix progress animation spawn entry bug
The ring progress has been decoupled from the entry spawn animation and is done via code and animation curves instead.
This way we can play the rest of the spawn animation while cancelling our ring spawn earlier than intended. When users pressed restart then play immediately, it would take a second for the ring progression to match the actual timer progress since it was coupled with the spawn animation (after the spawn animation completed, the ring progression would "snap" back to the correct place). Now, when the user presses the play button, the ring progression will match the timer with the cost of the ring progression animation being cancelled a bit earlier, however like I said, this will allow the ring progression to match the timer correctly.