Fix media creator tick animation bug
Screenshot had cut off timer values due to being mid animation. Tick animation probably got invoked by on value change since it occurs on text change. So we stopped turning on the animation when we want to play a tick animation and moved the animation enable to a method and used that in the places that need it.