Fix caret run-time null reference bug
When we switch pages, we re-generate/intialize the digit format. However the TMP input field caret only appears once the gameobject is active. When we init the digit format, we don't have the digit format active because we are animating out of the previous page (but not on first launch). The init is where we attach the caret reference, which doesn't quite exist until the timer page shows. So we simply hook up the reference when we show the timer page/when the page becomes visible.