Implement isHoldable property to ClickButton's to prevent ClickButton OnClick spam
When a button is marked as holdable it can be held down and invoked multiple times in a single hold/press. Otherwise the button can only be activated once for pressing down, then once for releasing...I.e. It won't invoke multiple OnClick events when the user is holding the button.