Refactor SYSTEM settings file saving so that most recent version is used and backported to the other location (remote/local) as fallback.
Basically we've rewritten how we load our SYSTEM settings files from either local or remote storage.

First thing we do when attempting to load a SYSTEM settings file is determine which path is valid/the most recent save location. See the MostRecentSaveLocation enum for all details on potential outcomes.

We then take the most recent save file and backport it to the other directory or location: meaning if the most recent save is a Steam cloud save, we load it and overwrite the local storage save with the Steam cloud save. The inverse is also true, meaning if the most recent save is the local storage save, we load it and overwrite the Steam cloud save with the local storage save.

The intent here is to have a fallback in both cases. So if the user has a steam cloud save but no local save (happens when you install the app on a new computer with the same steam account), we take the steam cloud save and store it locally. This way if the internet does go down later or steam is not present anymore, we can fall back and utilize the local save.

Now we could have programmed the app to check the steam userdata directory instead of creating a copy of the remote data locally but this way if the user deletes Steam and wants to keep this app on their computer without steam they can still have local saves working without the need for a Steam directory being present on the machine. Probably a very rare case, but what is done is done. Maybe we can revisit this in the future...