Improve custom audio support to better keep track of added/removed files (WIP)
There is still a bug present related to the alarm bank dictionary keys not matching or corresponding with the dropdown option index value. Moving forward, we'll need a way for the dropdown options to be associated with a key not just a specific index. I can explain this a bit better...

Basically, when a dropdown option get's added, it gets added at the next available count index. So if we have a alarm bank of 4 preinstalled alarms [0,1,2,3], the next alarm will get set to index [4]. Then if we add another dropdown option, that alarm gets set to index [5]. Now, if we try to remove an option dropdown, we will remove the first custom alarm which was assigned an index of [4]. However because our dropdown options is a list, it cascades...meaning the second alarm option we added has been set to a key of 5 but because the previous option was removed, it's alarm index is now set to a [4]. Which is not going to be assigned to anything in the dictionary, since key 4 no longer exists when the first custom was removed.

I'm going to assume we want to add the key value to the dropdown option name. That way we can identify which alarm bank dictionary key to look for instead of relying soley on the dropdown option index.