Only consume long break pomodoros/tomatoes when long break timer is complete
Previously we would consume the tomatoes when the user pressed play on the long break. But since users can dispose of their pomodoros via trash if they restart/cancel/interrupt the long break, it's not really needed to dispose on play. This way the user can restart their long breaks without having to re-accumulate pomodoro and also the side benefit of making labels work with long breaks.

2 birds, 1 stone.